+++ commons-spring-test/src/test/java/com/hb0730/commons/spring/test/IpUtilsTest.java
@@ -1,0 +1,31 @@
+package com.hb0730.commons.spring.test;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+import static org.junit.Assert.*;
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringBootTest
+public class IpUtilsTest {
+    private MockMvc mockMvc;
+    @Autowired
+    private WebApplicationContext context;
+
+    @Before
+    public void init() {
+        mockMvc = MockMvcBuilders.webAppContextSetup(context).build();
+    }
+
+
+    @Test
+    public void getIpTest() {
+
+    }
+}